+++ src/PSDocs/Models/DocumentNodeType.cs
@@ -1,0 +1,17 @@
+﻿namespace PSDocs.Models
+{
+    public enum DocumentNodeType
+    {
+        Document,
+
+        Section,
+
+        Table,
+
+        Code,
+
+        Note,
+
+        Warning
+    }
+}