+++ tests/PSDocs.Tests/PSDocs.Table.Tests.ps1
@@ -71,7 +71,7 @@ Describe 'PSDocs -- Table keyword' {
         }
 
         It 'Should match expected format' {
-            Get-Content -Path $outputDoc -Raw | Should match '\|LICENSE\|False\|(\n|\r){1,2}\|README.md\|False\|';
+            Get-Content -Path $outputDoc -Raw | Should -Match '\|LICENSE\|False\|(\n|\r){1,2}\|README.md\|False\|';
         }
     }
 
@@ -110,11 +110,46 @@ Describe 'PSDocs -- Table keyword' {
         Invoke-PSDocument -Name 'TableWithNull' -InputObject @{ ResourceType = @{  } } -OutputPath $outputPath;
 
         It 'Should have generated output' {
-            Test-Path -Path $outputDoc | Should be $True;
+            Test-Path -Path $outputDoc | Should -Be $True;
         }
 
         It 'Should match expected format' {
-            Get-Content -Path $outputDoc -Raw | Should match '(## Windows features\r\n)$';
+            Get-Content -Path $outputDoc -Raw | Should -Match '(## Windows features\r\n)$';
+        }
+    }
+
+    Context 'Table with multiline column' {
+
+        $testObject = [PSCustomObject]@{
+            Name = 'Test'
+            Description = "This is a`r`ndescription`r`nsplit`r`nover`r`nmultiple`r`nlines."
+        }
+
+        # Define a test document with a multiple column in a table
+        document 'TableWithMultilineColumn' {
+            $testObject | Table;
+        }
+
+        $outputDoc = "$outputPath\TableWithMultilineColumn.md";
+        Invoke-PSDocument -Name 'TableWithMultilineColumn' -OutputPath $outputPath;
+
+        It 'Should have generated output' {
+            Test-Path -Path $outputDoc | Should -Be $True;
+        }
+
+        It 'Should match expected format' {
+            Get-Content -Path $outputDoc -Raw | Should -Match 'This is a description split over multiple lines\.';
+        }
+
+        $option = New-PSDocumentOption @{
+            'Markdown.WrapSeparator' = '<br />'
+        }
+
+        $outputDoc = "$outputPath\TableWithMultilineColumnCustom.md";
+        Invoke-PSDocument -Name 'TableWithMultilineColumn' -InstanceName 'TableWithMultilineColumnCustom' -OutputPath $outputPath -Option $option;
+
+        It 'Should use wrap separator' {
+            Get-Content -Path $outputDoc -Raw | Should -Match 'This is a\<br /\>description\<br /\>split\<br /\>over\<br /\>multiple\<br /\>lines\.';
         }
     }
 }